requirements.txt
h5py
thop
torchsummaryX
sklearn